Update app directory structure
