+++ app/assets/style/style.sass
@@ -1,0 +1,1 @@
+body
