+++ app/components/App.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Provider } from 'react-redux';
-import App from './containers/App'
-import { configureStore } from './store';
+import App from '../modules/App/containers/App'
+import { configureStore } from '../store';
 
 export default () => (
   <Provider store={configureStore()}>
