+++ app/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { AppContainer } from 'react-hot-loader';
-import App from './App';
+import App from './components/App';
 
 const appEl = document.getElementById('app');
 const render = Component => {
@@ -14,7 +14,7 @@ const render = Component => {
 };
 
 render(App);
-if (module.hot) module.hot.accept('./App', () => {
-  const newApp = require('./App').default;
+if (module.hot) module.hot.accept('./components/App', () => {
+  const newApp = require('./components/App').default;
   render(newApp);
 });
