+++ webpack.config.js
@@ -41,6 +41,6 @@ module.exports = {
     host: 'localhost',
     port: 3000,
     historyApiFallback: true,
-    hot: true,
+    hot: true
   }
 };
